# travo/backend/docs/api_spec.yaml
openapi: 3.0.0
info:
  title: TRAVO API
  description: |
    API for the TRAVO travel application, providing services for monument identification,
    travel recommendations, crowd prediction, and AI assistant features.
  version: 1.0.0
  contact:
    name: TRAVO Team
    email: support@travo.example.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.travo.example.com
    description: Production server
...
paths:
...
  /assistant/ask:
    post:
      tags:
        - assistant
      summary: Ask the AI assistant
      description: Send a text query to get information about monuments, history, or visiting tips
      operationId: askAssistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantQuery'
      responses:
        '200':
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Invalid query type or empty query
  /assistant/voice_to_text:
    post:
      tags:
        - assistant
      summary: Convert voice to text
      description: Convert voice audio (base64 encoded) to text
      operationId: voiceToText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceToTextRequest'
      responses:
        '200':
          description: Voice to text conversion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceToTextResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Invalid audio format or empty audio data
  /assistant/text_to_voice:
    post:
      tags:
        - assistant
      summary: Convert text to voice
      description: Convert text to voice audio (base64 encoded)
      operationId: textToVoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextToVoiceRequest'
      responses:
        '200':
          description: Text to voice conversion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextToVoiceResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Empty text or invalid voice parameters
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    QueryType:
      type: string
      enum: [MONUMENT_INFO, HISTORICAL_FACT, VISITING_TIP, GENERAL]
      example: MONUMENT_INFO
    AssistantQuery:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          example: "Tell me about the Sphinx"
        query_type:
          $ref: '#/components/schemas/QueryType'
        monument_id:
          type: string
          example: "sphinx_giza"
    AssistantResponse:
      type: object
      properties:
        response:
          type: string
          example: "The Great Sphinx of Giza is a limestone statue of a reclining sphinx, a mythical creature with the head of a human and the body of a lion. It stands on the Giza Plateau on the west bank of the Nile in Giza, Egypt. It is the oldest known monumental sculpture in Egypt and is commonly believed to have been built by ancient Egyptians of the Old Kingdom during the reign of the Pharaoh Khafre (c. 2558–2532 BC)."
        query_type:
          $ref: '#/components/schemas/QueryType'
        source:
          type: string
          enum: [AI, RULE_BASED]
          example: AI
    VoiceToTextRequest:
      type: object
      required:
        - audio_data
      properties:
        audio_data:
          type: string
          format: byte
          description: Base64 encoded audio data
        audio_format:
          type: string
          enum: [WAV, MP3, OGG]
          default: WAV
          example: WAV
    VoiceToTextResponse:
      type: object
      properties:
        text:
          type: string
          example: "Tell me about the Sphinx"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.92
    TextToVoiceRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          example: "The Great Sphinx of Giza is a limestone statue of a reclining sphinx."
        voice_type:
          type: string
          enum: [MALE, FEMALE]
          default: FEMALE
          example: FEMALE
        language_code:
          type: string
          default: en-US
          example: en-US
    TextToVoiceResponse:
      type: object
      properties:
        audio_data:
          type: string
          format: byte
          description: Base64 encoded audio data
        audio_format:
          type: string
          enum: [WAV, MP3]
          example: MP3
        duration_seconds:
          type: number
          format: float
          example: 5.2